+++ controllers/MailerController.php
@@ -94,58 +94,6 @@ class MailerController extends Controller
      * Lists of all sending emails.
      * @return mixed
      */
-    /*public function actionIndex()
-    {
-
-        $data = [];
-        $mailsPath = $this->module->mailsPath;
-        $mailParser = new \ZBateson\MailMimeParser\MailMimeParser();
-        $dir = \yii\helpers\BaseFileHelper::normalizePath(Yii::getAlias($mailsPath));
-        $emls = \yii\helpers\BaseFileHelper::findFiles($dir, [
-            'only' => ['*.eml']
-        ]);
-
-        foreach ($emls as $eml) {
-
-            if (strpos($eml, $dir) !== 0)
-                throw new \Exception("Something wrong: {$eml}\n");
-
-            $sourcePath = \yii\helpers\BaseFileHelper::normalizePath($eml);
-            $rawEml = fopen($sourcePath, 'r');
-            $message = $mailParser->parse($rawEml);
-            fclose($rawEml);
-
-            $data[] = [
-                'message_id' => $message->getHeaderValue('Message-ID'),
-                'datetime' => $message->getHeaderValue('Date'),
-                'subject' => $message->getHeaderValue('Subject'),
-                'email_from' => $message->getHeaderValue('From'),
-                'email_to' => $message->getHeaderValue('To'),
-                'email_copy' => $message->getHeaderValue('CC'),
-                'mime_version' => $message->getHeaderValue('MIME-Version'),
-                'html_content' => $message->getHtmlContent(),
-                'html_content' => \yii\helpers\HtmlPurifier::process($message->getHtmlContent()),
-                'text_content' => $message->getTextContent(),
-                'filename' => basename($sourcePath),
-                'source' => $sourcePath
-            ];
-
-        }
-
-        $dataProvider = new \yii\data\ArrayDataProvider([
-            'allModels' => $data,
-            'pagination' => [
-                'pageSize' => 10,
-            ],
-            'sort' => [
-                'attributes' => ['id', 'datetime', 'subject', 'email_from', 'email_to'],
-            ],
-        ]);
-
-        return $this->render('index', [
-            'dataProvider' => $dataProvider
-        ]);
-    }*/
     public function actionIndex()
     {
         $searchModel = new MailsSearch();
@@ -208,6 +156,14 @@ class MailerController extends Controller
             ]);
 
         return $this->goBack(['index']);
+    }
+
+    public function actionClear()
+    {
+        $model = new Mails();
+        $model->deleteAll();
+
+        return $this->redirect(['mailer/index']);
     }
 
     /**